backend/pyproject.toml
[project]
name = "agent-control-plane-backend"
version = "0.1.0"
description = "FastAPI WebSocket backend for the Agent Control Plane demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-httplib2>=0.2.0",
]

[project.scripts]
serve = "uvicorn:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
